Fix hydration error
